[feat] migrate to Astro v5 Content Layer API
- Move content config from src/content/config.ts to src/content.config.ts
- Add glob loaders for posts and slides collections
- Update render() function usage in dynamic routes
- Replace entry.slug references with entry.id
- Verify build compatibility with new Content Layer API
